+++ pages/[slug].js
@@ -5,55 +5,62 @@ import Head from 'next/head'
 import { useTina } from 'tinacms/dist/react'
 import { TinaMarkdown } from 'tinacms/dist/rich-text'
 import client from '../tina/__generated__/client'
+import animationStyles from "@/styles/animations.module.scss";
+import postStyles from "@/styles/posts.module.scss";
+import Layout from '@/components/layout';
 
-const BlogPage = (props) => {
+const Page = (props) => {
   const { data } = useTina({
     query: props.query,
     variables: props.variables,
     data: props.data,
   })
 
-  // console.log(data.post.body)
+  function MinsRead() {
+    const wc = data.post.wordCount;
+    if (wc <= 360) {
+      return "1 min reading time";
+    }
+    return `${Math.floor(wc / 180)} min reading time`;
+  }
+
+  // console.log('page data:', data.post.body.children[0].value)
+
   return (
-    <>
-      <Head>
-        {/* Tailwind CDN */}
-        <link
-          rel="stylesheet"
-          href="https://cdnjs.cloudflare.com/ajax/libs/tailwindcss/2.2.7/tailwind.min.css"
-          integrity="sha512-y6ZMKFUQrn+UUEVoqYe8ApScqbjuhjqzTuwUMEGMDuhS2niI8KA3vhH2LenreqJXQS+iIXVTRL2iaNfJbDNA1Q=="
-          crossOrigin="anonymous"
-          referrerPolicy="no-referrer"
-        />
-      </Head>
-      <div>
-        <div
-          style={{
-            textAlign: 'center',
-          }}
-        >
-          <h1 className="text-3xl m-8 text-center leading-8 font-extrabold tracking-tight text-gray-900 sm:text-4xl">
-            {data.post.title}
-          </h1>
-          <ContentSection content={data.post.body}></ContentSection>
-          <duv dangerouslySetInnerHTML={{ __html: data.post.body.children }}></duv>
-        </div>
-        <div className="bg-green-100 text-center">
-          Lost and looking for a place to start?
-          <a
-            href="https://tina.io/guides/tina-cloud/getting-started/overview/"
-            className="text-blue-500 underline"
-          >
-            {' '}
-            Check out this guide
-          </a>{' '}
-          to see how add TinaCMS to an existing Next.js site.
+    <Layout post title={data.post.title}>
+      <div className={animationStyles.cssanimation}>
+        <div className={postStyles["content-title"]}>
+          <h1 className={postStyles.post_title}>{data.post.title}</h1>
         </div>
       </div>
-    </>
-  )
+
+      <h3>/ {data.post.contributor}</h3>
+      <h4 className={postStyles.meta}>{data.post.tags.join(", ")} &mdash; <MinsRead /></h4>
+      {data.post.featured ? (<h4 className={postStyles.gold}>No. {data.post.category}</h4>) : null}
+
+
+      <article id="cr-article" className={postStyles["#cr-article"]}>
+        <OPMHTML content={data.post.body.children} />
+      </article>
+
+      <div className={postStyles["copyright-footer"]}>
+        <h4>This work belongs to {data.post.contributor} © <br /> Copy, reproduction, and modification are not permitted without permission</h4>
+      </div>
+
+      <div className={postStyles["post-nav"]}>
+        <div className={postStyles["post-nav-prev"]}>
+          {data.post.prevPost ? (<h4><Link href={data.post.prevPost}>&lt;prev</Link></h4>) : null}
+        </div>
+        <div className={postStyles["post-nav-next"]}>
+          {data.post.nextPost ? (<h4><Link href={data.post.nextPost}>next&gt;</Link></h4>) : null}
+        </div>
+      </div>
+
+    </Layout>
+  );
 }
 
+export default Page
 
 export const getStaticProps = async ({ params }) => {
   let data = {}
@@ -82,139 +89,44 @@ export const getStaticPaths = async () => {
   const postsListData = await client.queries.postConnection()
 
   return {
-    paths: postsListData.data.postConnection.edges.map((post) => ({
-      params: { slug: post.node._sys.filename },
-    })),
+    paths: postsListData.data.postConnection.edges.map((post) => (
+      {
+        params: { slug: post.node._sys.filename },
+      })),
     fallback: false,
   }
 }
 
-export default BlogPage
-
-const PageSection = (props) => {
-  return (
-    <>
-      <h2>{props.heading}</h2>
-      <p>{props.content}</p>
-    </>
-  )
+/**
+ * 
+ * @param {[AST]} content An array of AST nodes from the rich text editor
+ * @returns Parsed raw output from the rich text editor
+ */
+const OPMHTML = ({ content }) => {
+  return content.map((astChild) => {
+    console.log(astChild, astChild.type)
+    return (astChild.type === 'html') ? (
+      <div dangerouslySetInnerHTML={{ __html: astChild.value }} />
+    ) : (
+      recurseAsHTML(astChild)
+    )
+  });
 }
 
-const components = {
-  PageSection: PageSection,
-}
+/**
+ * 
+ * @param {AST} ast Abstract Syntax Tree nodes from the rich text editor
+ * @returns A <p> element of the HTML-md styled text one-level down.  
+ */
+const recurseAsHTML = (ast) => {
+  let text = [];
+  ast.children.map((child) => {
+    child.type === 'html_inline' || child.type === 'html' ? (
+      text.push(child.value)
+    ) : (
+      text.push(child.text)
+    )
+  });
 
-const ContentSection = ({ content }) => {
-  return (
-    <div className="relative py-16 bg-white overflow-hidden">
-      <div className="hidden lg:block lg:absolute lg:inset-y-0 lg:h-full lg:w-full">
-        <div
-          className="relative h-full text-lg max-w-prose mx-auto"
-          aria-hidden="true"
-        >
-          <svg
-            className="absolute top-12 left-full transform translate-x-32"
-            width={404}
-            height={384}
-            fill="none"
-            viewBox="0 0 404 384"
-          >
-            <defs>
-              <pattern
-                id="74b3fd99-0a6f-4271-bef2-e80eeafdf357"
-                x={0}
-                y={0}
-                width={20}
-                height={20}
-                patternUnits="userSpaceOnUse"
-              >
-                <rect
-                  x={0}
-                  y={0}
-                  width={4}
-                  height={4}
-                  className="text-gray-200"
-                  fill="currentColor"
-                />
-              </pattern>
-            </defs>
-            <rect
-              width={404}
-              height={384}
-              fill="url(#74b3fd99-0a6f-4271-bef2-e80eeafdf357)"
-            />
-          </svg>
-          <svg
-            className="absolute top-1/2 right-full transform -translate-y-1/2 -translate-x-32"
-            width={404}
-            height={384}
-            fill="none"
-            viewBox="0 0 404 384"
-          >
-            <defs>
-              <pattern
-                id="f210dbf6-a58d-4871-961e-36d5016a0f49"
-                x={0}
-                y={0}
-                width={20}
-                height={20}
-                patternUnits="userSpaceOnUse"
-              >
-                <rect
-                  x={0}
-                  y={0}
-                  width={4}
-                  height={4}
-                  className="text-gray-200"
-                  fill="currentColor"
-                />
-              </pattern>
-            </defs>
-            <rect
-              width={404}
-              height={384}
-              fill="url(#f210dbf6-a58d-4871-961e-36d5016a0f49)"
-            />
-          </svg>
-          <svg
-            className="absolute bottom-12 left-full transform translate-x-32"
-            width={404}
-            height={384}
-            fill="none"
-            viewBox="0 0 404 384"
-          >
-            <defs>
-              <pattern
-                id="d3eb07ae-5182-43e6-857d-35c643af9034"
-                x={0}
-                y={0}
-                width={20}
-                height={20}
-                patternUnits="userSpaceOnUse"
-              >
-                <rect
-                  x={0}
-                  y={0}
-                  width={4}
-                  height={4}
-                  className="text-gray-200"
-                  fill="currentColor"
-                />
-              </pattern>
-            </defs>
-            <rect
-              width={404}
-              height={384}
-              fill="url(#d3eb07ae-5182-43e6-857d-35c643af9034)"
-            />
-          </svg>
-        </div>
-      </div>
-      <div className="relative px-4 sm:px-6 lg:px-8">
-        <div className="text-lg max-w-prose mx-auto">
-          <TinaMarkdown components={components} content={content} />
-        </div>
-      </div>
-    </div>
-  )
+  return <p dangerouslySetInnerHTML={{ __html: text.join('') }} />;
 }